+++ app/src/main/java/com/xsolla/android/storesdkexample/ui/vm/base/ViewModelFactory.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModelProvider
 import com.xsolla.android.storesdkexample.ui.vm.VmProfile
 
 class ViewModelFactory(private val params: Any) : ViewModelProvider.Factory {
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(VmProfile::class.java)) {
             return VmProfile(params as Resources) as T
         }