+++ xsolla-inventory-sdk/src/main/java/com/xsolla/android/inventory/util/EngineUtils.kt
@@ -1,0 +1,6 @@
+package com.xsolla.android.inventory.util
+
+object EngineUtils {
+    @JvmStatic
+    var engineSpec = ""
+}