+++ xsolla-login-sdk/src/main/java/com/xsolla/android/login/util/EngineUtils.kt
@@ -1,0 +1,6 @@
+package com.xsolla.android.login.util
+
+object EngineUtils {
+    @JvmStatic
+    var engineSpec = ""
+}