+++ xsolla-store-sdk/src/main/java/com/xsolla/android/store/XStore.kt
@@ -28,6 +28,7 @@ import com.xsolla.android.store.entity.response.order.WsOrderResponse
 import com.xsolla.android.store.entity.response.payment.CreateOrderByVirtualCurrencyResponse
 import com.xsolla.android.store.entity.response.payment.CreateOrderResponse
 import com.xsolla.android.store.entity.response.payment.CreatePaymentTokenResponse
+import com.xsolla.android.store.util.EngineUtils
 import okhttp3.*
 import org.json.JSONObject
 import retrofit2.Call
@@ -74,6 +75,7 @@ class XStore private constructor(
                     .addHeader("X-ENGINE-V", Build.VERSION.RELEASE)
                     .addHeader("X-SDK", "STORE")
                     .addHeader("X-SDK-V", BuildConfig.VERSION_NAME)
+                    .addHeader("X-GAMEENGINE-SPEC", EngineUtils.engineSpec)
                     .url(
                         originalRequest.url().newBuilder()
                             .addQueryParameter("engine", "android")
@@ -471,7 +473,7 @@ class XStore private constructor(
         fun createOrderFromCartById(
             callback: CreateOrderCallback,
             cartId: String,
-            options: PaymentOptions? = null
+            options: PaymentOptions? = PaymentOptions()
         ) {
             val body = CreateCartOrderRequestBody(options)
             getInstance().storeApi.createOrderFromCartById(getInstance().projectId, cartId, body)
@@ -509,7 +511,7 @@ class XStore private constructor(
         @JvmOverloads
         fun createOrderFromCurrentCart(
             callback: CreateOrderCallback,
-            options: PaymentOptions? = null
+            options: PaymentOptions? = PaymentOptions()
         ) {
             val body = CreateCartOrderRequestBody(options)
             getInstance().storeApi.createOrderFromCurrentCart(getInstance().projectId, body)
@@ -550,7 +552,7 @@ class XStore private constructor(
         fun createOrderByItemSku(
             callback: CreateOrderCallback,
             itemSku: String,
-            options: PaymentOptions? = null,
+            options: PaymentOptions? = PaymentOptions(),
             quantity: Long = 1
         ) {
             val body = CreateSkuOrderRequestBody(quantity, options)